Refactor env vars
